refactor: split proxy engine out of main.go into cohesive files
main.go was 1620 lines mixing CLI/bootstrap with the proxy engine.
Move engine symbols into cohesive same-package files (config.go,
dial.go, headers.go, response.go, errors_proxy.go, connect.go,
direct.go, forward.go, version.go). Pure symbol relocation: no
behaviour, flag, exit-code, or log-format change; the in-package
test suite passes unchanged.
